doubleからfloatに変更
doubleからfloatに変更

+++ db/migrate/20210905091913_create_cooks.rb
@@ -11,8 +11,8 @@ class CreateCooks < ActiveRecord::Migration[6.0]
       t.string :open_day, null:false
       t.string :holiday_day, null:false
       t.string :regular_holiday, null:false
-      t.double(8,6) :lat,null: false
-      t.double(9,6) :lng,null: false
+      t.float :lat,null: false
+      t.float :lng,null: false
